Remove deprecated XA/Direct3D9/IntelCL state trackers
The XA state tracker and Direct3D9 state trackers were both
marked as deprecated and disabled by default in 25.1 and
have alreaday been removed in the main devel branch so are
going to be removed in 25.2 scheduled for August timeframe
so disable and obsolete them now to prepare for the cutoff.
The Intel CL support was also removed in the 25.2 cycle.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>